find_all is called like so:
	searched_node ->find_all (matching, output_list);

So after "this->match (in)", we should push_back THIS, not IN, since IN is the node we're trying to match, not a node in the collection we're searching in.

This fixes a bug where the above call returned a list containing multiple copies of the first parameter to find_all.


git-svn-id: http://maketea.googlecode.com/svn/trunk@94 e7869359-c92b-0410-9105-cbbd88f55910